Rename RxWebSocket to WalletLinkWebSocket
