+++ packages/wallet-sdk/src/connection/WalletLinkWebSocket.ts
